Sign the first state when joining a channel.
While incorporating virtual-funding into the create-and-fund workflow, I
ran into the issue of validating the app definition and data before
supporting the first state. It turns out that there's no benefit to
validating the first state at a later point -- the only code capable of
validating the first state is the application itself. Therefore, when
the app calls JOIN_CHANNEL, the wallet should assume an acceptable first
state, and might as well sign it straight away.