Refactor application states (again)
This is a simplification of the application protocol based on a discussion with Alex Gap,
where he helped me realize that the workflow should listen to
JOIN_CHANNEL _right away_ in case the app is quick to reply.

https://xstate.js.org/viz/?gist=d884fdb11a028b1ee2b924f3fb832108

The main difference is we _don't_ sign the first state when we receive
the JOIN_CHANNEL event. It seems reasonable to do it this -- consider the four cases:
1. I (Alice) am creating the channel
  a. funding it directly
  b. funding it virtually
2. I (Alice) joining the channel
  a. funding it directly
  b. funding it virtually

Case 1a:
  - I now know my opponent's wallet will use direct funding, since it's
  part of the OpenChannel message
  - Bob's wallet will support the preFundSetup AFTER Bob confirms in the
  UI, which is AFTER Bob's wallet called JOIN_CHANNEL
  - Thus, when I receive a supported state, I know everything's ok on
  Bob's end
Case 1b:
  i) I now know my opponent's wallet will use virtual funding
  ii) We'll both support preFundSetup right away
  iii) We'll both attempt to reserve funds concurrently
  iv) If we both succeed, we'll continue with virtual funding
  v) If one of us fails, the other simply has to release the funds

TODO: Further work can (should?)
I) rename createAndFund to just fundApp
II) pull the createAndFund workflow into the application workflow

If we do (II), then we can do (ii) and (iii) in parallel states,
quickening the virtual-funding setup by a one-way trip.