ChannelStoreData/Entry does not have a signatures field
Instead, the signatures are stored along with the state variables and
the hash.

The type T = Record<string, S> type is dangerous, beacuse if
if I have a signatureStore: T variable, then the compiler will thin that
signatureStore['whatever random string I want'] has type S.

This happened to me in the browser, leading to a runtime error that
could have been prevented by stronger typings.

Since we don't want to store signatures without the associated state
variables anyway, we might as well store them with the state variables.